requirements.txt
numpy~=1.19.0
pandas~=1.0.5
scipy~=1.5.1
pylint~=2.5.3
keras~=2.4.3
sklearn~=0.0
Flask~=1.1.1
markdown~=3.2.1
pygments~=2.5.2
jsonpickle~=1.3
pymongo~=3.4.0
requests~=2.24.0
tensorflow~=1.12.0
jsonschema~=3.2.0
pyopenssl~=19.1.0
python-dotenv~=0.14.0
statsmodels~=0.12.0
tqdm~=4.51.0